+++ lib/src/sc_builder.dart
@@ -40,6 +40,7 @@ class ScBuilder {
 
   /// Applies the builder to this scope and all its children
   void applyToScope(Scope scope, {bool applyToChildren = true}) {
+    bluePrint.onInstantiate(hostScope: scope);
     inserts.applyToScope(scope);
     nodeReplacer.applyToScope(scope);
     nodeAdder.applyToScope(scope);