+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Add tests for applying builders to nodes created by builders
 - Add shouldStopProcessingAfter to optimize application of builders
 - Add Scope.ls to print all children and nodes of a scope
+- Add onInstantiate bluePrint
 
 ### Changed